fix error larastan to resolve the ci eror

+++ app/Models/SubmissionReviewer.php
@@ -46,6 +46,7 @@ class SubmissionReviewer extends Model
     /**
      * @return HasMany<ReviewerFormAssignment, $this>
      */
+    /** @return HasMany<ReviewerFormAssignment, $this> */
     public function reviewerFormAssignments(): HasMany
     {
         return $this->hasMany(ReviewerFormAssignment::class);
@@ -156,7 +157,7 @@ class SubmissionReviewer extends Model
 
         $completedAssignments = $this->completedReviewerFormAssignments()->count();
 
-        return round(($completedAssignments / $totalAssignments) * 100);
+        return (int) round(($completedAssignments / $totalAssignments) * 100);
     }
 
     // NEW: Update evaluation status based on current form completion
